feat: Additional edits for TextSearchView
